Comment out dom.storage.enabled
This is a feature that more and more sites utilize these days and most
of them break horribly without it and on top of that, don't even bother
to check whether it's available or not :(

Now we'll just rely on private browsing mode to clear it when the
browser is closed. Just like cookies.