Bigger update to fit with HTTPSE and NoScript
Since we're recommend (and using) NoScript and HTTPS Everywhere, we can just harden some more stuff from the beginning. 

HTTPS Everywhere:
* Enable all popup messages to notify on observatory problems/changes/news
* Enable the toolbar hint (not annoying)

NoScript:
The settings are to fit with Firefox, so e.g. we doesn't want to allow webgl (except on whitelistening pages the user sets - if not set also apply these restrictions to whitelist sites too -> must be manually set because it's up to the user).

* Enable ABE (don't do this on Tor)
* Enable STS (don't do this on Tor)
* Don't allow the cliuck-to-play stuff, it's depcreated and only for older FF versions (or maybe ESR variants of it) 
* Only allow the current tab on change (not all opened tabs)
* Cascade the menu, it's a visual tweak 
* Show some other visual options to not get useless things displayed and reveal only stuff that is important, uBlock or uMatrix already do a better finer job for this 
* Changed the default policy to harden possible security related 'leaks' 
* Media is enabled since this breaks so many pages (due .js, HTML5) [use Random Agent Spoofer or uMarix for this]