added update hardening
Biggest weakness I can see it still the about:config update change problem in FF so we ensure we use all the latest url's. Of course if you decided to no use e.g. peerconnection it makes less sense but not everyone disabled it so give them the ability to use the latest secured servers and update resources.