No need to completely disable WebRTC anymore
No need to disable WebRTC anymore, because WebRTC is not overall bad and some pages will be not working anymore if you disabled it competently. Now we got an option to use it without leaking the IP, because now only the IP will be used that is visible to the server.