Replace "Smart HTTPS" with "HTTPS by default"
Smart HTTPS seems to break a few sites in non-obvious ways and its
defaults are not great. "HTTPS by default" has none of these
problems and just does one thing well: it changes the default
scheme when typing scheme-less URLs in the address bar.

Also add the trimURLs setting to user.js since it's recommended
by the author of "HTTPS by default" and it will probably appeal to
the user.js crowd too.